Kafka-Consumer/requirements.txt
scikit-learn==0.24.2
kafka-python==2.0.1
pandas
category_encoders